move PropagationStyle to top level (aesthetics)
