+++ dd-trace-api/src/main/java/datadog/trace/api/PropagationStyle.java
@@ -1,0 +1,7 @@
+package datadog.trace.api;
+
+public enum PropagationStyle {
+  DATADOG,
+  B3,
+  HAYSTACK
+}
