+++ dd-trace-core/src/test/groovy/datadog/trace/core/propagation/HttpExtractorTest.groovy
@@ -5,8 +5,8 @@ import datadog.trace.api.DDId
 import datadog.trace.util.test.DDSpecification
 import spock.lang.Shared
 
-import static datadog.trace.api.Config.PropagationStyle.B3
-import static datadog.trace.api.Config.PropagationStyle.DATADOG
+import static datadog.trace.api.PropagationStyle.B3
+import static datadog.trace.api.PropagationStyle.DATADOG
 import static datadog.trace.core.CoreTracer.TRACE_ID_MAX
 
 class HttpExtractorTest extends DDSpecification {
